@@ -19,9 +19,9 @@
     }
   ],
   "require": {
-    "php": ">7.2.0",
+    "php": ">8.0.0",
     "psr/http-client": "^1.0",
-    "toolkit/stdlib": "^1.0"
+    "toolkit/stdlib": "^2.0"
   },
   "autoload": {
     "psr-4": {
