@@ -17,7 +17,7 @@ $namespaces = [
 
 spl_autoload_register(static function ($class) use ($namespaces): void {
     foreach ($namespaces as $np => $dir) {
-        if (0 === strpos($class, $np)) {
+        if (str_starts_with($class, $np)) {
             $path = str_replace('\\', '/', substr($class, strlen($np)));
             $file = $dir . "/$path.php";
 
